auto2/navi_poc/requirements.txt
pynput>=1.7
pyautogui>=0.9
pillow>=10.0
pytesseract>=0.3.10
pywin32>=306
psutil>=5.9
ollama>=0.1.0
